brain_tumor_api/requirements.txt

torch
torchvision
Pillow
scikit-learn
matplotlib
captum
nibabel
